+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increases unit-tests coverage of core components
 - Implements the Bernoulli distribution in the PRG class
 - Plugins can now be reloaded
-- Adds button to maximize graph widgets
-- Adds button to pause all experiments
+- Projects page: Adds button to maximize graph widgets
+- Projects page: Adds button to pause all experiments
+- Experiments table: Allows hiding columns
 - Graph widgets: Allows scaling the nodes and edges
 - Graph widgets: Allows setting the edge with
 - Graph widgets: Allows exporting nodes to file